requirements.txt
amqp==5.2.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==5.0.1
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.13.4
billiard==4.2.1
bleach==6.1.0
boto3==1.34.84
botocore==1.34.162
Brotli==1.1.0
bs4==0.0.2
cachetools==5.5.0
celery==5.4.0
certifi==2025.4.26
cffi==1.17.1
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cron-descriptor==1.4.5
cryptography==43.0.3
cssselect2==0.7.0
curl_cffi==0.11.1
cycler==0.12.1
debugpy==1.8.1
decorator==5.2.1
defusedxml==0.7.1
diff-match-patch==20241021
distlib==0.3.8
dj-database-url==2.3.0
Django==5.2.1
django-allauth==64.2.0
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==25.1
django-heroku==0.3.1
django-import-export==4.3.1
django-querycount==0.8.3
django-storages==1.14.2
django-timezone-field==7.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
dnspython==1.16.0
drf-yasg==1.21.10
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.2.0
Faker==25.3.0
fastjsonschema==2.19.1
filelock==3.13.4
fonttools==4.54.1
fqdn==1.5.1
frozendict==2.4.6
gevent==24.11.1
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
identity==0.5.0
idna==3.10
inflection==0.5.1
ipykernel==6.29.4
ipython==8.36.0
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
kagglehub==0.3.6
kiwisolver==1.4.7
kombu==5.4.2
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
ms-identity-python @ git+https://github.com/azure-samples/ms-identity-python@bc793f365c1fd8567e1dec5f18064dd67bbb3afb
msal==1.31.0
mssql-django==1.5
multitasking==0.0.11
mysql-connector-python==9.2.0
mysqlclient==2.2.7
narwhals==1.39.1
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==2.1.3
oauthlib==3.2.2
openpyxl==3.1.5
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
peewee==3.18.1
pillow==11.0.0
platformdirs==4.2.0
plotly==6.1.0
prometheus_client==0.20.0
prompt_toolkit==3.0.51
proto-plus==1.25.0
protobuf==5.29.2
psutil==5.9.8
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.8.0
pymongo==3.12.0
PyMuPDF==1.25.1
PyMySQL==1.1.1
pyodbc==5.1.0
pyparsing==3.2.0
PyPDF2==3.0.1
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-json-logger==2.0.7
python3-openid==3.2.0
pytube==15.0.0
pytz==2024.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.2
qrcode==8.0
razorpay==1.4.2
redis==6.0.0
referencing==0.35.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
resend==2.4.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.4
scikit-learn==1.6.0
scipy==1.15.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==60.2.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
statsmodels==0.14.4
tablib==3.7.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.4.0
uuid==1.30
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
weasyprint==64.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.37.1
whitenoise==6.9.0
XlsxWriter==3.2.0
yfinance==0.2.61
youtube-transcript-api==1.0.1
zope.event==5.0
zope.interface==7.2
zopfli==0.2.3.post1
